Create a new MsalSettings class for end-user to access the parameter settings for MSAL auth request.
